DJANGO/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0
django-dump-load-utf8==0.0.4
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
Jinja2==3.1.3
MarkupSafe==2.1.3
psycopg2==2.9.9
pycparser==2.21
sqlparse==0.4.4
tzdata==2023.3
Werkzeug==3.0.1